Finished Keyboard input for the player icon for now.
1 thing bugging me, but given the nature of how i'll set it up for pad,
it's fine (moving forward in coRoutine, then press d, and waits for next
loop of it. just a small delay that's hard to get used to.)

This is kinda outdone by the fact i'll make the analouge movement a
free-Roam style movement, and snap to the closest node when there is no
more analouge input.  and the dPad will functoin like the WASD movement,
but the dpad movement will be more for small movements rather than
constantly using it, generally speaking, so it should be fine.